Give HttpReads traits consistent names, remove HttpErrorFunctions where it's not needed
