+++ src/main/scala/uk/gov/hmrc/play/http/reads/HtmlHttpReads.scala
@@ -22,6 +22,6 @@ trait HtmlHttpReads {
   def bodyToHtml = HttpReads[Html] { (method, url, response) => Html(response.body) }
 
   //TODO this shouldn't have to be a def, look at initialisation order to solve
-  implicit def readToHtml: HttpReads[Html] = ErrorReads.convertFailuresToExceptions or bodyToHtml
+  implicit def readToHtml: HttpReads[Html] = ErrorHttpReads.convertFailuresToExceptions or bodyToHtml
 }
 object HtmlHttpReads extends HtmlHttpReads
